Add API for user

+++ server/app/models/Submission.scala
@@ -3,8 +3,8 @@ package models
 import org.joda.time.DateTime
 
 case class Submission(
-  id:             Int,
-  idQuestion:     Int,
+  id:             Long,
+  idQuestion:     Long,
   user:           User,
   release:        DateTime,
   fileName:       String,
